+++ R/predict_survival.R
@@ -5,10 +5,9 @@
 #' @export predict_survival.model_fit
 #' @export
 predict_survival.model_fit <-
-  function(object, new_data, .time, ...) {
+  function(object, new_data, time, ...) {
 
-    if (is.null(object$spec$method$pred$survival))
-      rlang::abort("No survival prediction method defined for this engine.")
+    check_spec_pred_type(object, "survival")
 
     if (inherits(object$fit, "try-error")) {
       rlang::warn("Model fit failed; cannot make predictions.")
@@ -22,7 +21,7 @@ predict_survival.model_fit <-
       new_data <- object$spec$method$pred$survival$pre(new_data, object)
 
     # Pass some extra arguments to be used in post-processor
-    object$spec$method$pred$survival$args$.time <- .time
+    object$spec$method$pred$survival$args$time <- time
     pred_call <- make_pred_call(object$spec$method$pred$survival)
 
     res <- eval_tidy(pred_call)
@@ -35,9 +34,9 @@ predict_survival.model_fit <-
     res
   }
 
-# @export
-# @keywords internal
-# @rdname other_predict
-# @inheritParams predict.model_fit
+#' @export
+#' @keywords internal
+#' @rdname other_predict
+#' @inheritParams predict_survival.model_fit
 predict_survival <- function (object, ...)
   UseMethod("predict_survival")