+++ R/svm_linear_data.R
@@ -61,7 +61,7 @@ set_encoding(
     predictor_indicators = "none",
     compute_intercept = FALSE,
     remove_intercept = FALSE,
-    allow_sparse_x = FALSE
+    allow_sparse_x = TRUE
   )
 )
 
@@ -73,7 +73,7 @@ set_encoding(
     predictor_indicators = "none",
     compute_intercept = FALSE,
     remove_intercept = FALSE,
-    allow_sparse_x = FALSE
+    allow_sparse_x = TRUE
   )
 )
 
@@ -115,27 +115,6 @@ set_pred(
   value = list(
     pre = NULL,
     post = svm_linear_post,
-    func = c(fun = "predict"),
-    args =
-      list(
-        object = quote(object$fit),
-        newx = expr(as.matrix(new_data))
-      )
-  )
-)
-set_pred(
-  model = "svm_linear",
-  eng = "LiblineaR",
-  mode = "classification",
-  type = "prob",
-  value = list(
-    pre = function(x, object) {
-      rlang::abort(
-        paste0("The LiblineaR engine does not support class probabilities ",
-               "for any `svm` models.")
-      )
-    },
-    post = NULL,
     func = c(fun = "predict"),
     args =
       list(