+++ tests/testthat/test_svm_linear.R
@@ -104,8 +104,8 @@ test_that('updating', {
 })
 
 test_that('bad input', {
-  expect_error(svm_linear(mode = "reallyunknown"))
   expect_error(translate(svm_linear(mode = "regression") %>% set_engine( NULL)))
+  expect_error(svm_linear(mode = "reallyunknown"))
   expect_error(translate(svm_linear(mode = "regression") %>% set_engine("LiblineaR", type = 3)))
   expect_error(translate(svm_linear(mode = "classification") %>% set_engine("LiblineaR", type = 11)))
 })
@@ -142,6 +142,13 @@ test_that('linear svm regression: LiblineaR', {
   expect_false(has_multi_predict(res))
   expect_equal(multi_predict_args(res), NA_character_)
   expect_output(print(res), "parsnip model object")
+
+  expect_error(
+    tidy_res <- tidy(res),
+    NA
+  )
+  expect_s3_class(tidy_res, "tbl_df")
+  expect_equal(colnames(tidy_res), c("term", "estimate"))
 
   expect_error(
     fit(
@@ -273,12 +280,12 @@ test_that('linear svm classification prediction: LiblineaR', {
 
   expect_error(
     predict(cls_form, hpc_no_m[ind, -5], type = "prob"),
-    "The LiblineaR engine does not support class probabilities"
+    "No prob prediction method available for this model"
   )
 
   expect_error(
     predict(cls_xy_form, hpc_no_m[ind, -5], type = "prob"),
-    "The LiblineaR engine does not support class probabilities"
+    "No prob prediction method available for this model"
   )
 
 })